requirements.txt
streamlit
langchain
langchain-openai
langchain-community
beautifulsoup4
python-dotenv
chromadb
langchain_experimental
tabulate